Фикс разбиения на файлы
