--- api/domain/services/tools/UserServiceExceptionHandler.py
@@ -1,11 +1,0 @@
-from shared.exceptions.UserServiceExceptions import UserAlreadyExistsException, UserNotExistsException
-
-
-class UserServiceExceptionHandler:
-    @staticmethod
-    def handle(exc: Exception):
-        if isinstance(exc, UserAlreadyExistsException):
-            raise exc
-
-        elif isinstance(exc, UserNotExistsException):
-            raise exc
