--- api/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-testpaths =
-    user_component_tests.py
