+++ src/common/exceptions/InfrastructureExceptions.py
