+++ src/plugins/user/services/tools/UserServiceExceptionHandler.py
@@ -1,0 +1,11 @@
+from src.common.exceptions.UserServiceExceptions import UserAlreadyExistsException, UserNotExistsException
+
+
+class UserServiceExceptionHandler:
+    @staticmethod
+    def handle(exc: Exception):
+        if isinstance(exc, UserAlreadyExistsException):
+            raise exc
+
+        elif isinstance(exc, UserNotExistsException):
+            raise exc
